Reorganise packages in admin client tests
